# hello/hello.swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "hello/hello.proto",
    "version": "version not set"
  },
  "schemes": [
    "http",
    "https"
  ],
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
    "/v1/example/echo/{id}": {
      "post": {
        "operationId": "Echo",
        "responses": {
          "200": {
            "description": "",
            "schema": {
              "$ref": "#/definitions/helloSimpleMessage"
            }
          }
        },
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "tags": [
          "HelloService"
        ]
      }
    },
    "/v1/example/echo/{id}/{num}": {
      "get": {
        "operationId": "Echo",
        "responses": {
          "200": {
            "description": "",
            "schema": {
              "$ref": "#/definitions/helloSimpleMessage"
            }
          }
        },
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "num",
            "in": "path",
            "required": true,
            "type": "string",
            "format": "int64"
          }
        ],
        "tags": [
          "HelloService"
        ]
      }
    },
    "/v1/example/echo_post": {
      "post": {
        "operationId": "EchoPost",
        "responses": {
          "200": {
            "description": "",
            "schema": {
              "$ref": "#/definitions/helloStringMessage"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/helloStringMessage"
            }
          }
        ],
        "tags": [
          "HelloService"
        ]
      }
    }
  },
  "definitions": {
    "helloSimpleMessage": {
      "type": "object",
      "properties": {
        "id": {
          "type": "string"
        },
        "num": {
          "type": "string",
          "format": "int64"
        }
      }
    },
    "helloStringMessage": {
      "type": "object",
      "properties": {
        "value": {
          "type": "string"
        }
      }
    }
  }
}